Langsdale on the Go! - 2015-03-26
This is the updated version of the LotG! mobile interface. It reflects
a number of changes since the original 2013 version of the web app.
Most notably EDS search and a link to eprint.ubalt.edu from the main
screen; as well as updates to the maps, location and book locator to
reflect our temporary location during renovations of the Langsdale
building.